refactor: modularize server setup
