+++ Dockerfile
@@ -8,4 +8,4 @@ RUN pip install -e .
 EXPOSE 8000
 HEALTHCHECK --interval=30s --timeout=5s --start-period=5s \
     CMD curl -f http://localhost:8000/health || exit 1
-CMD ["bash", "-c", "python scripts/init_db_schema.py && uvicorn main:app --host 0.0.0.0 --port 8000"]
+CMD ["bash", "-c", "python scripts/init_db_schema.py && uvicorn main:create_app --factory --host 0.0.0.0 --port 8000"]
